Increased debug output precision. Fixed rounding bug on timeline when finding a clips actual frame number. Fixed bug with a clips JSON properties output.

+++ include/WriterBase.h
@@ -91,12 +91,12 @@ namespace openshot
 		void AppendDebugItem(Json::Value debug_item);
 
 		/// Append debug information as JSON
-		void AppendDebugMethod(string method_name, string arg1_name, int arg1_value,
-												   string arg2_name, int arg2_value,
-												   string arg3_name, int arg3_value,
-												   string arg4_name, int arg4_value,
-												   string arg5_name, int arg5_value,
-												   string arg6_name, int arg6_value);
+		void AppendDebugMethod(string method_name, string arg1_name, float arg1_value,
+												   string arg2_name, float arg2_value,
+												   string arg3_name, float arg3_value,
+												   string arg4_name, float arg4_value,
+												   string arg5_name, float arg5_value,
+												   string arg6_name, float arg6_value);
 
 	public:
 		/// Constructor for WriterBase class, many things are initialized here
